no more welcome view/controller. Home is now the home page.
